add hooks for ci configurations into Makefile.main
Adds two new targets: `ci-install` and `ci-script`. These are meant to be the
only targets running in the install and test phases both on `.travis.yml` and
`appveyor.yml`.

The relevant `.travis.yml` section would look like:

```yaml
install:
  - make ci-install
script:
  - make ci-script
```

And `appveyor.yml` would be:

```yaml
install:
  - set PATH=%GOPATH%\bin;C:\go\bin;C:\mingw-w64\x86_64-7.2.0-posix-seh-rt_v5-rev1\mingw64\bin;C:\msys64\usr\bin;%PATH%;"C:\Program Files\Git\mingw64\bin"
  - mingw32-make ci-install
test_script:
  - mingw32-make ci-script
```

The behaviour of these rules is as follows:

* `ci-install` always runs `dependencies`.

* `ci-script` runs `test-coverage`, `codecov` and `packages`.

This commit also changes the example `.travis.yml` to use multi-stage builds, which offer
more flexibility to define build matrices, conditional stages and so on. The deploy section
did not receive any significant change yet, but it might be changed to a `ci-*`-style rule
too in a further change to simplify the CI definition.

Signed-off-by: Santiago M. Mola <santi@mola.io>